+++ lib/project_types/theme/commands/push.rb
@@ -72,6 +72,7 @@ module Theme
               @ctx.done(@ctx.message("theme.push.done", theme.preview_url, theme.editor_url))
             end
           end
+          raise ShopifyCLI::AbortSilent if syncer.has_any_error?
         rescue ShopifyCLI::API::APIRequestNotFoundError
           @ctx.abort(@ctx.message("theme.push.theme_not_found", theme.id))
         ensure