+++ lib/shopify_cli/theme/syncer/operation.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ShopifyCLI
+  module Theme
+    class Syncer
+      class Operation < Struct.new(:method, :file)
+        def to_s
+          "#{method} #{file&.relative_path}"
+        end
+      end
+    end
+  end
+end