+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 From version 2.6.0, the sections in this file adhere to the [keep a changelog](https://keepachangelog.com/en/1.0.0/) specification.
 ## [Unreleased]
+### Fixed
 * [#1763](https://github.com/Shopify/shopify-cli/pull/1722): Fix: Tunnel --PORT parameter not working in Node.js app.
+* [#1769](https://github.com/Shopify/shopify-cli/pull/1769): Fix `theme push --development --json` to output the proper exit code
 
 ## Version 2.7.1
 ### Fixed